Add Av1CoefDecoderGpu - GPU AV1 token entropy decoder (CUDA + CPU bit-exact)
Symmetric companion to Av1CoefEncoderGpu. 100% ILGPU port of
Av1CoefDecoder.ReadCoeffsTxb for the v1 keyframe configurations
(Tx8x8/Tx16x16, DCT_DCT, DC_PRED). Mirrors every step of libaom
read_coeffs_txb: txb_skip, tx_type, eob_pt, eob_extra,
coeff_base_eob, coeff_base, coeff_lps (BR loop), dc_sign + raw signs
+ golomb tail + dequant + clamp.

Caller pre-supplies dequant scales (qDc + qAc), so the helper does
not need to pack the DcQuantQtx / AcQuantQtx tables. v1 always uses
8-bit (txScale = 0 for both Tx8x8 / Tx16x16).

Round-trip test driver Av1CoefRoundTripKernel encodes via
Av1CoefEncoderGpu then decodes via Av1CoefDecoderGpu in the same
dispatch. Five tests verify the decoder produces input coefs
verbatim when qDc=qAc=1: AllZero / DcOnly / NegativeDc /
RandomSparse / HighMagnitude (golomb tail).

PMT cross-backend results: 11/15 PASS.
- CUDA: 5/5 PASS
- CPU: 5/5 PASS
- OpenCL: 1/5 PASS, 4/5 FAIL

Bug isolated to the multi-symbol CDF round-trip on OpenCL.
Av1CdfRoundTripKernel is a minimal repro (encode 16 4-sym CDF
symbols, decode them back); it fails 1/3 backends (OpenCL only) with
sym[1] decoded as 0 instead of 1. The binary-only EncodeBoolQ15 /
DecodeBoolQ15 path works on all 3 backends (already exercised by the
shipping Av1RangeCoderGpu round-trip test); the bug is specific to
the multi-sym EncodeCdfQ15 / DecodeCdfQ15 path on the OpenCL
backend. Per Rule 2 (Fix Libraries First), this is being filed with
Geordi as a SpawnDev.ILGPU OpenCL backend issue rather than worked
around in Codecs.

+++ SpawnDev.Codecs/Video/Av1/Av1CoefEncoderGpu.cs
@@ -303,11 +303,13 @@ public static class Av1CoefEncoderGpu
         ArrayView<byte> levels, long levelsBase)
     {
         // Zero the buffer first - libaom zeroes the levels buffer
-        // before the per-coef writes.
+        // before the per-coef writes. Buffer size fits in int (max
+        // ~456 bytes for Tx16x16); use int counter for OpenCL backend
+        // codegen safety.
         int paddedStride = (1 << bhl) + Av1TxbCommonGpu.TxPadHor;
         int paddedRows = height + Av1TxbCommonGpu.TxPadVer;
-        long bufBytes = (long)paddedStride * paddedRows + Av1TxbCommonGpu.TxPadEnd;
-        for (long i = 0; i < bufBytes; i++) levels[levelsBase + i] = 0;
+        int bufBytes = paddedStride * paddedRows + Av1TxbCommonGpu.TxPadEnd;
+        for (int i = 0; i < bufBytes; i++) levels[levelsBase + i] = 0;
 
         int levelsOff = Av1TxbCommonGpu.SetLevelsOffset(height);
         for (int col = 0; col < width; col++)
